+++ client/src/index.tsx
@@ -89,6 +89,10 @@ const Codes10Management = React.lazy(() => import("./pages/admin/management/10-c
 const Add10Code = React.lazy(() => import("./pages/admin/management/10-codes/add-code"));
 const Edit10Code = React.lazy(() => import("./pages/admin/management/10-codes/edit-code"));
 
+const PenalCodesManagement = React.lazy(() => import("./pages/admin/management/penal-codes"));
+const AddPenalCode = React.lazy(() => import("./pages/admin/management/penal-codes/add-code"));
+const EditPenalCode = React.lazy(() => import("./pages/admin/management/penal-codes/edit-code"));
+
 ReactDOM.render(
   <React.StrictMode>
     <Redux.Provider store={store}>
@@ -205,6 +209,22 @@ ReactDOM.render(
               path="/admin/manage/10-codes"
               requirement="admin"
               Component={Codes10Management}
+            />
+
+            <AuthRoute
+              path="/admin/manage/penal-codes/edit/:id"
+              requirement="admin"
+              Component={EditPenalCode}
+            />
+            <AuthRoute
+              path="/admin/manage/penal-codes/add"
+              requirement="admin"
+              Component={AddPenalCode}
+            />
+            <AuthRoute
+              path="/admin/manage/penal-codes"
+              requirement="admin"
+              Component={PenalCodesManagement}
             />
 
             <AuthRoute