+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 
 **If you need help updating/installing or have found any bugs, please join [our Discord server](https://discord.gg/eGnrPqEH7U) or open a [GitHub issue here](https://github.com/Dev-CasperTheGhost/snaily-cadv3/issues/new/choose)**
 
-## 1.2.5
+## 1.3.0
 
 - Add version check in client side
-- Server side rendering (No more CORS errors!)
-- Customize penal codes and 10 codes
+- Major: Server side rendering (No more CORS errors!)
+- Major: Customize penal codes and 10 codes
 
 ## 1.2.4